requirements.txt
certifi @ file:///C:/ci/certifi_1655907258560/work/certifi
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
requests==2.32.2
urllib3==2.2.1
wincertstore==0.2
